renaming NL to make them default analyses
