+++ scripts/lib/template_ds001_AFNI_level1
@@ -7,19 +7,24 @@
 afni_proc.py -subj_id ${subj}                                                  \
         -script proc.${subj} -scr_overwrite                                    \
         -blocks tshift align tlrc volreg blur mask scale regress               \
-        -copy_anat $anat_dir/${sub}_T1w.nii                                    \
+        -copy_anat $anat_dir/anatSS.${sub}.nii                                 \
+		-anat_has_skull no					       \
         -tcat_remove_first_trs 2                                               \
         -align_opts_aea -cost mi -ginormous_move -check_flip                   \
         -dsets                                                                 \
             $func_dir/${sub}_task-balloonanalogrisktask_run-01_bold.nii.gz     \
             $func_dir/${sub}_task-balloonanalogrisktask_run-02_bold.nii.gz     \
             $func_dir/${sub}_task-balloonanalogrisktask_run-03_bold.nii.gz     \
-	-anat_uniform_method unifize                                               \
-        -tlrc_base MNI_avg152T1+tlrc                                           \
+        -tlrc_base MNI152_2009_template.nii.gz                                 \
         -volreg_warp_dxyz 2                                                    \
         -volreg_align_to third                                                 \
         -volreg_align_e2a                                                      \
         -volreg_tlrc_warp                                                      \
+	-tlrc_NL_warp							       \
+	-tlrc_NL_warped_dsets						       \
+		$anat_dir/anatQQ.${sub}.nii				       \
+		$anat_dir/anatQQ.${sub}.aff12.1D			       \
+		$anat_dir/anatQQ.${sub}_WARP.nii			       \
         -blur_size 5.0                                                         \
         -regress_stim_times                                                    \
             $stim_dir/${sub}_combined_cash_demean_afni.1d                      \
@@ -38,11 +43,11 @@ afni_proc.py -subj_id ${subj}                                                  \
         'BLOCK(0.772,1)' 'BLOCK(0.772,1)' 'dmBLOCK'                            \
         -regress_stim_types                                                    \
             AM2 AM1 AM2 AM1 AM2 AM2 AM1                                        \
-        -regress_censor_motion 0.3                                             \
+	-regress_3dD_stop						       \
+	-regress_reml_exec 						       \
         -regress_opts_3dD                                                      \
-            -gltsym 'SYM: pumps_demean -control_pumps_demean'                  \
+            -gltsym 'SYM: pumps_demean[1] -control_pumps_demean[1]'            \
         -glt_label 1 pumps_demean_vs_ctrl_demean                               \
         -regress_make_ideal_sum sum_ideal.1D                                   \
         -regress_est_blur_epits                                                \
         -regress_est_blur_errts
-
