--- scripts/lib/template_ds001_AFNI_level1_NL
@@ -1,53 +1,0 @@
-#!/usr/bin/env tcsh
-
-# created by uber_subject.py: version 0.39 (March 21, 2016)
-# creation date: Tue Sep 13 15:59:22 2016
-
-# run afni_proc.py to create a single subject processing script
-afni_proc.py -subj_id ${subj}                                                  \
-        -script proc.${subj} -scr_overwrite                                    \
-        -blocks tshift align tlrc volreg blur mask scale regress               \
-        -copy_anat $anat_dir/anatSS.${sub}.nii                                 \
-		-anat_has_skull no					       \
-        -tcat_remove_first_trs 2                                               \
-        -align_opts_aea -cost mi -ginormous_move -check_flip                   \
-        -dsets                                                                 \
-            $func_dir/${sub}_task-balloonanalogrisktask_run-01_bold.nii.gz     \
-            $func_dir/${sub}_task-balloonanalogrisktask_run-02_bold.nii.gz     \
-            $func_dir/${sub}_task-balloonanalogrisktask_run-03_bold.nii.gz     \
-        -tlrc_base MNI152_2009_template.nii.gz                                 \
-        -volreg_warp_dxyz 2                                                    \
-        -volreg_align_to third                                                 \
-        -volreg_align_e2a                                                      \
-        -volreg_tlrc_warp                                                      \
-	-tlrc_NL_warp							       \
-	-tlrc_NL_warped_dsets						       \
-		$anat_dir/anatQQ.${sub}.nii				       \
-		$anat_dir/anatQQ.${sub}.aff12.1D			       \
-		$anat_dir/anatQQ.${sub}_WARP.nii			       \
-        -blur_size 5.0                                                         \
-        -regress_stim_times                                                    \
-            $stim_dir/${sub}_combined_cash_demean_afni.1d                      \
-            $stim_dir/${sub}_combined_cash_RT_afni.1d                          \
-            $stim_dir/${sub}_combined_control_pumps_demean_afni.1d             \
-            $stim_dir/${sub}_combined_control_pumps_RT_afni.1d                 \
-            $stim_dir/${sub}_combined_explode_demean_afni.1d                   \
-            $stim_dir/${sub}_combined_pumps_demean_afni.1d                     \
-            $stim_dir/${sub}_combined_pumps_RT_afni.1d                         \
-        -regress_stim_labels                                                   \
-            cash_demean cash_RT control_pumps_demean                           \
-            control_pumps_RT explode_demean pumps_demean                       \
-            pumps_RT                                                           \
-        -regress_basis_multi                                                   \
-            'BLOCK(0.772,1)' 'dmBLOCK' 'BLOCK(0.772,1)' 'dmBLOCK'              \
-        'BLOCK(0.772,1)' 'BLOCK(0.772,1)' 'dmBLOCK'                            \
-        -regress_stim_types                                                    \
-            AM2 AM1 AM2 AM1 AM2 AM2 AM1                                        \
-	-regress_3dD_stop						       \
-	-regress_reml_exec 						       \
-        -regress_opts_3dD                                                      \
-            -gltsym 'SYM: pumps_demean[1] -control_pumps_demean[1]'            \
-        -glt_label 1 pumps_demean_vs_ctrl_demean                               \
-        -regress_make_ideal_sum sum_ideal.1D                                   \
-        -regress_est_blur_epits                                                \
-        -regress_est_blur_errts
