+++ scripts/lib/template_ds120_AFNI_level1
@@ -7,7 +7,8 @@
 afni_proc.py -subj_id ${subj}                                                  \
         -script proc.${subj} -scr_overwrite                                    \
         -blocks tshift align tlrc volreg blur mask scale regress               \
-        -copy_anat $anat_dir/${sub}_T1w.nii                                    \
+        -copy_anat $anat_dir/anatSS.${sub}.nii                             \
+		-anat_has_skull no					       \
         -tcat_remove_first_trs 4                                               \
         -align_opts_aea -giant_move -check_flip                                \
         -dsets                                                                 \
@@ -15,11 +16,16 @@ afni_proc.py -subj_id ${subj}                                                  \
             $func_dir/${sub}_task-antisaccadetaskwithfixedorder_run-02_bold.nii.gz    \
             $func_dir/${sub}_task-antisaccadetaskwithfixedorder_run-03_bold.nii.gz    \
         -tshift_opts_ts -tpattern alt+z                                        \
-        -tlrc_base MNI_avg152T1+tlrc                                           \
+        -tlrc_base MNI152_2009_template.nii.gz                                 \
         -volreg_warp_dxyz 2                                                    \
         -volreg_align_to third                                                 \
         -volreg_align_e2a                                                      \
         -volreg_tlrc_warp                                                      \
+	-tlrc_NL_warp							       \
+	-tlrc_NL_warped_dsets						       \
+		$anat_dir/anatQQ.${sub}.nii				       \
+		$anat_dir/anatQQ.${sub}.aff12.1D			       \
+		$anat_dir/anatQQ.${sub}_WARP.nii			       \
         -blur_size 5.0                                                         \
         -regress_stim_times                                                    \
             $stim_dir/${sub}_combined_neutral_afni.1d                          \
@@ -28,7 +34,8 @@ afni_proc.py -subj_id ${subj}                                                  \
             neutral reward                                                     \
         -regress_basis_multi                                                   \
             'SIN(0,24,8)' 'SIN(0,24,8)'                                        \
-        -regress_censor_motion 0.3                                             \
+	-regress_3dD_stop						       \
+	-regress_reml_exec 						       \
         -regress_opts_3dD                                                      \
             -gltsym 'SYM: neutral'                                             \
         -glt_label 1 neutral_vs_baseline                                       \
