+++ scripts/process_ds001_AFNI.py
@@ -7,6 +7,7 @@ from lib.afni_processing import run_subject_level_analyses
 from lib.afni_processing import run_group_level_analysis
 from lib.afni_processing import run_permutation_test
 from lib.afni_processing import mean_mni_images
+from lib.afni_processing import run_SSWarper
 
 raw_dir = '/home/maullz/NIDM-Ex/BIDS_Data/DATA/BIDS/ds001_R2.0.4'
 
@@ -42,7 +43,7 @@ cwd = os.path.dirname(os.path.realpath(__file__))
 
 # Copy raw anatomical and functional data to the preprocessing directory and
 # run BET on the anatomical images
-copy_raw(raw_dir, preproc_dir)
+#copy_raw(raw_dir, preproc_dir)
 
 # Directory to store the onset files
 onset_dir = os.path.join(afni_dir, 'ONSETS')
@@ -62,20 +63,24 @@ conditions = (
     ('control_pumps_RT', ('control_pumps_demean', 'response_time')))
 
 # Create onset files based on BIDS tsv files
-cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time)
+#cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time)
 
+SSWarper_template = os.path.join(cwd, 'lib', 'template_AFNI_SSWarper')
 sub_level_template = os.path.join(cwd, 'lib', 'template_ds001_AFNI_level1')
 grp_level_template = os.path.join(cwd, 'lib', 'template_ds001_AFNI_level2')
 perm_template = os.path.join(cwd, 'lib', 'template_ds001_AFNI_perm_test')
 
+# Run SSWarper AFNI command on each subject to strip skull and warp to MNI template
+#run_SSWarper(preproc_dir, SSWarper_template)
+
 # Run a GLM combining all the fMRI runs of each subject
-run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
+#run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
 
 # Run the group-level GLM
-run_group_level_analysis(level1_dir, level2_dir, grp_level_template)
+#run_group_level_analysis(level1_dir, level2_dir, grp_level_template)
 
 # Run a permutation test
-#run_permutation_test(level1_dir, perm_dir, perm_template)
+run_permutation_test(level1_dir, perm_dir, perm_template)
 
 # Create mean and standard deviations maps of the mean func and anat images in MNI space
 #mean_mni_images(preproc_dir, level1_dir, mni_dir)
