+++ scripts/process_ds001_FSL.py
@@ -6,7 +6,6 @@ from lib.fsl_processing import run_subject_level_analyses
 from lib.fsl_processing import run_group_level_analysis
 from lib.fsl_processing import run_permutation_test
 from lib.fsl_processing import mean_mni_images
-from lib.fsl_processing import nidm_export
 
 raw_dir = '/storage/essicd/data/NIDM-Ex/BIDS_Data/DATA/BIDS/ds001_R2.0.4'
 results_dir = '/storage/essicd/data/NIDM-Ex/BIDS_Data/RESULTS/SOFTWARE_COMPARISON/ds001'
@@ -33,7 +32,7 @@ cwd = os.path.dirname(os.path.realpath(__file__))
 
 # Copy raw anatomical and functional data to the preprocessing directory and
 # run BET on the anatomical images
-copy_and_BET(raw_dir, preproc_dir)
+#copy_and_BET(raw_dir, preproc_dir)
 
 # Directory to store the onset files
 onsetDir = os.path.join(fsl_dir, 'ONSETS')
@@ -50,7 +49,7 @@ conditions = (
     ('control_pumps_RT', ('control_pumps_demean', 'response_time')))
 
 # Create 3-columns onset files based on BIDS tsv files
-cond_files = create_fsl_onset_files(raw_dir, onsetDir, conditions, removed_TR_time)
+#cond_files = create_fsl_onset_files(raw_dir, onsetDir, conditions, removed_TR_time)
 
 run_level_fsf = os.path.join(cwd, 'lib', 'template_ds001_FSL_level1.fsf')
 sub_level_fsf = os.path.join(cwd, 'lib', 'template_ds001_FSL_level2.fsf')
@@ -58,19 +57,16 @@ grp_level_fsf = os.path.join(cwd, 'lib', 'template_ds001_FSL_level3.fsf')
 perm_template = os.path.join(cwd, 'lib', 'template_ds001_FSL_perm_test')
 
 # Run a GLM for each fMRI run of each subject
-run_run_level_analyses(preproc_dir, run_level_fsf, level1_dir, cond_files)
+#run_run_level_analyses(preproc_dir, run_level_fsf, level1_dir, cond_files)
 
 # Run a GLM combining all the fMRI runs of each subject
-run_subject_level_analyses(level1_dir, sub_level_fsf, level2_dir)
+#run_subject_level_analyses(level1_dir, sub_level_fsf, level2_dir)
 
 # Run the group-level GLM
-run_group_level_analysis(level2_dir, grp_level_fsf, level3_dir, '1')
-
-# Export group-level results as a NIDM-Results pack
-nidm_export(level1_dir, level3_dir)
+#run_group_level_analysis(level2_dir, grp_level_fsf, level3_dir, '1')
 
 # Run a permutation test
-#run_permutation_test(level1_dir, perm_dir, perm_template)
+run_permutation_test(level1_dir, perm_dir, perm_template)
 
 # Create mean and standard deviations maps of the mean func and anat images in MNI space
 #mean_mni_images(preproc_dir, level1_dir, mni_dir)
