+++ scripts/process_ds109_AFNI.py
@@ -7,6 +7,7 @@ from lib.afni_processing import run_subject_level_analyses
 from lib.afni_processing import run_group_level_analysis
 from lib.afni_processing import run_permutation_test
 from lib.afni_processing import mean_mni_images
+from lib.afni_processing import run_SSWarper
 
 raw_dir = '/home/maullz/NIDM-Ex/BIDS_Data/DATA/BIDS/ds000109_R2.0.1'
 
@@ -46,7 +47,7 @@ cwd = os.path.dirname(os.path.realpath(__file__))
 
 # Copy raw anatomical and functional data to the preprocessing directory and
 # run BET on the anatomical images
-copy_raw(raw_dir, preproc_dir, subject_ids)
+#copy_raw(raw_dir, preproc_dir, subject_ids)
 
 # Directory to store the onset files
 onset_dir = os.path.join(afni_dir, 'ONSETS')
@@ -62,20 +63,24 @@ conditions = (
     ('false_photo_question', ('false photo question', 'duration')))
 
 # Create onset files based on BIDS tsv files
-cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time, subject_ids)
+#cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time, subject_ids)
 
+SSWarper_template = os.path.join(cwd, 'lib', 'template_AFNI_SSWarper')
 sub_level_template = os.path.join(cwd, 'lib', 'template_ds109_AFNI_level1')
 grp_level_template = os.path.join(cwd, 'lib', 'template_ds109_AFNI_level2')
 perm_template = os.path.join(cwd, 'lib', 'template_ds109_AFNI_perm_test')
 
+# Run SSWarper AFNI command on each subject to strip skull and warp to MNI template
+#run_SSWarper(preproc_dir, SSWarper_template)
+
 # Run a GLM combining all the fMRI runs of each subject
-run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
+#run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
 
 # Run the group-level GLM
-run_group_level_analysis(level1_dir, level2_dir, grp_level_template)
+#run_group_level_analysis(level1_dir, level2_dir, grp_level_template)
 
 # Run a permutation test
-#run_permutation_test(level1_dir, perm_dir, perm_template)
+run_permutation_test(level1_dir, perm_dir, perm_template)
 
 # Create mean and standard deviations maps of the mean func and anat images in MNI space
 #mean_mni_images(preproc_dir, level1_dir, mni_dir)
