+++ scripts/process_ds120_AFNI.py
@@ -7,7 +7,9 @@ from subprocess import check_call
 from lib.afni_processing import copy_raw, create_afni_onset_files
 from lib.afni_processing import run_subject_level_analyses
 from lib.afni_processing import run_group_level_analysis
+from lib.afni_processing import run_permutation_test
 from lib.afni_processing import mean_mni_images
+from lib.afni_processing import run_SSWarper
 
 pre_raw_dir = '/home/maullz/NIDM-Ex/BIDS_Data/DATA/BIDS/ds120_R1.0.0'
 results_dir = '/home/maullz/NIDM-Ex/BIDS_Data/RESULTS/SOFTWARE_COMPARISON/ds120'
@@ -47,7 +49,7 @@ cwd = os.path.dirname(os.path.realpath(__file__))
 
 # Copy raw anatomical and functional data to the preprocessing directory and
 # run BET on the anatomical images
-copy_raw(raw_dir, preproc_dir, subject_ids)
+#copy_raw(raw_dir, preproc_dir, subject_ids)
 
 # Directory to store the onset files
 onset_dir = os.path.join(afni_dir, 'ONSETS')
@@ -61,13 +63,17 @@ conditions = (
     ('reward', ('reward_resp', 'duration')))
 
 # Create onset files based on BIDS tsv files
-cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time, subject_ids)
+#cond_files = create_afni_onset_files(raw_dir, onset_dir, conditions, removed_TR_time, subject_ids)
 
+SSWarper_template = os.path.join(cwd, 'lib', 'template_AFNI_SSWarper')
 sub_level_template = os.path.join(cwd, 'lib', 'template_ds120_AFNI_level1')
 grp_level_template = os.path.join(cwd, 'lib', 'template_ds120_AFNI_level2')
 
+# Run SSWarper AFNI command on each subject to strip skull and warp to MNI template
+#run_SSWarper(preproc_dir, SSWarper_template)
+
 # Run a GLM combining all the fMRI runs of each subject
-run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
+#run_subject_level_analyses(preproc_dir, onset_dir, level1_dir, sub_level_template)
 
 # Run the group-level GLM
 run_group_level_analysis(level1_dir, level2_dir, grp_level_template)
